Remove dnsmock from test, implement NewTest, functional options, sort lifecycle into test.go
